backend/requirements.txt
# Django Core
django~=5.2.4

# REST Framework
djangorestframework~=3.15.0

# Firebase Integration
firebase-admin~=6.4.0

# CORS Support
django-cors-headers~=4.3.1

# HTTP Requests
requests~=2.32.3

# Development Tools
python-dotenv~=1.0.0